Issue 14. Restrictions put in to only allow those records with appropriate ties to the relevant website to be output over the services. In addition the ability to write to entities has been restricted to location, occurrence, occurrence_comment, person, sample, and user.
Outstanding issues:
Currently these restrictions apply to the top level model: we need to ensure that the use of superModels do not circumvent these restrictions - can these be used to access records on other websites?
Creation of a users_websites record when creating a user - how is this done and how is the site_role_id set?
Ensure that the users can not set the website_id explicitly unless creating a record, when the website_id must match that of the updating user.