Report which returns a list of squares, and the squares include a plot where the species has been recorded.
Also includes a report to populate the species parameter, although this is likely to need further adjustment to narrow its search.

The main report also includes a date before which occurrences are ignored. This is because the NPMS project is yearly, old data is read-only on the system and may need to be ignored in reports.