Fix possible incorrect record substatuses
If a record was verified as V1 or V2, then edited on a single record
form that had a hidden input setting occurrence:record_status to 'C'
then this forcing of the record status prevented the code firing on
the occurrence (or sample) model to correctly clean out the prior
verification status. This could result in incorrect status + substatus
combinations.