Added a can_edit column to occurrences associations list report
This tags records on the "left" side of associations as editable, as
those on the "right" side are entered into the same form so shouldn't be
accessed directly on the edit page. Therefore you can use the output of
this field to control the visibililty of edit actions in report grids.