Fix scheduled tasks being unable to write
A constraint to make database connections for reporting read-only was
removed. In the case where the 'report' and 'default' database
connections had identical connection strings and a process tried to use
both (as scheduled tasks does and possibly others) then the read-only
restriction could end up being applied to a connection where write
access is required.