Add child record grid tabs to samples and locations.
Add child record grid tabs to locations and samples.
Also adds asscoiated sample grid tab to locations, and links from child
record edit pages to their parents.
Includes a link from a sample to its location (if present) and a
location record autocomplete field for the sample.
GV views are updated to include the parent_id fields for samples and
locations, plus the location_id for the sample.
When creating a location from within the child grid, the parent is
automatically filled in.