Improves date/time information for redeterminations
* determinations.created_on now the time the determination was
envisaged by the user (even if not stored in the db yet)
* determinations.updated_on is the time the database was updated with
the determination record.

Because an occurrence can be saved without a determination and a
determination reflecting the original state of the record is only added
later when the record is redetermined, created_on and updated_on may be
different.