Report optimisations
To avoid possible bad abort early plans, now does an early count
limited to the first page of records to determine if less than that
number of records. If so, then alters the sort order to avoid using an
index scan which can be very slow if not enough records to fill the
page.