Multi-value overwrite fix
In old code, if you submitted a multi-value attribute value for a
sample twice, because the code only cleaned out different attribute
values, you end up with duplicates.
In the new code, each submission removes all the values that are not in
the current submission based on value ID. Therefore duplicates don’t
occur.