Re-organising of grid ref blurring functions.
Apart from a code tidy, the main change here is that the system used
for data entry is no longer used to default the output grid ref system.
Known areas of the world’s surface are still returned in the preferred
local projection. Other areas are always blurred using the appropriate
UTM strip and the output_sref is always a lat long, with a tilde prefix
if blurred. This fixes issues where a mix of projections can be seen on
the map for blurred grids and where web mercator was used for blurring
which gave a grid square size somewhat smaller than actually requested.