UKBMS Issue 119
The logic prevent the extra calculations is too restrictive, and seems
to cause issues when processing several samples on the same location at
the same time.
This extra check has been removed: it may mean that in a few cases the
summary calculations will be repeated, but from a performance viewpoint
there should also be a speedup as the queries are simplified for every
sample and occurrence.