Fix for Postgres13 compatibility
The query was getting columns for the entity from  both the warehouse
schema and the information_schema. In pg13 a 'name' datatype is used rather
than varchar for system info which is an unlisted sql_type. Rather than
add that type, maybe we were processing data we didn't need so the
information_schema columns are filtered out.
This may prove to be a mistake!