Passing 0 no longer causes unspecified parameter inclusion
Joins/wheres in report parameters which don't specify the operator
and value are included if a non-empty value is passed for the
parameter. 0 was being treated as non-empty. This caused UKBMS to
always remove sensitive sites from the annual summary even if admin.

Fixes https://github.com/BiologicalRecordsCentre/UKBMS-online/issues/345