Preprocessing standard parameters fixed
If preprocessing a parameter resulted in a change of datatype from string to integer
then when a value passed that returned nothing from the pre-processing query, the
replacement used to nullify the report results was of the wrong type causing an SQL
error. E.g. a filter on taxa_taxon_list_external_key_list=foo would crash the SQL.